refactor : JudgeService를 Judge 도메인으로 변경
