+++ src/test/java/service/JudgeTest.java
@@ -1,6 +1,7 @@
 package service;
 
 import domain.Cars;
+import domain.Judge;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,13 +11,13 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static util.Constants.MIN_FORWARD_NUMBER;
 
-class JudgeServiceTest {
+class JudgeTest {
 
-    private JudgeService judgeService;
+    private Judge judge;
 
     @BeforeEach
     void setUp() {
-        judgeService = new JudgeService();
+        judge = new Judge();
     }
 
     @DisplayName("우승자 1명 테스트")
@@ -28,7 +29,7 @@ class JudgeServiceTest {
         cars.getCars().get(1).drive(MIN_FORWARD_NUMBER);
         cars.getCars().get(1).drive(MIN_FORWARD_NUMBER); // "나"가 우승자
 
-        assertThat(judgeService.findWinners(cars)).containsAll(List.of("나"));
+        assertThat(judge.findWinners(cars)).containsAll(List.of("나"));
     }
 
     @DisplayName("우승자 여러명 테스트")
@@ -42,7 +43,7 @@ class JudgeServiceTest {
         cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
         cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
 
-        assertThat(judgeService.findWinners(cars)).containsAll(List.of("가", "다"));
+        assertThat(judge.findWinners(cars)).containsAll(List.of("가", "다"));
     }
 
     @DisplayName("모두가 우승자 테스트")
@@ -57,7 +58,7 @@ class JudgeServiceTest {
         cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
         cars.getCars().get(2).drive(MIN_FORWARD_NUMBER);
 
-        assertThat(judgeService.findWinners(cars)).containsAll(List.of("가", "나", "다"));
+        assertThat(judge.findWinners(cars)).containsAll(List.of("가", "나", "다"));
     }
 
 }
